Update of modules, files, functions names according to standard
