+++ app/urls.py
@@ -17,6 +17,6 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path('', include('Kryptonite.urls')),
+    path('', include('kryptonite.urls')),
     path('admin/', admin.site.urls)
 ]
