+++ kryptonite/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class KryptoniteConfig(AppConfig):
-    name = 'Kryptonite'
+    name = 'kryptonite'
