+++ kryptonite/dataservice/poloniex/models/ChartDataDto.py
