+++ kryptonite/dataservice/poloniex/models/TradeHistoryDto.py
