+++ kryptonite/static/external/sweetalert2_dark.css
