+++ kryptonite/static/utils.js
