+++ kryptonite/templates/_currency_buttons.html
