+++ kryptonite/templates/_exchange_dropdown.html
