+++ kryptonite/templates/_navigation.html
