+++ kryptonite/views.py
@@ -8,8 +8,8 @@ from django.http import JsonResponse
 from django.http import Http404
 from django.contrib import messages
 import sweetify
-from Kryptonite.DataService.Binance.BinanceClient import BinanceClient
-from Kryptonite.DataService.Poloniex.client import PoloniexClient
+from kryptonite.dataservice.binance.binance_client import BinanceClient
+from kryptonite.dataservice.poloniex.client import PoloniexClient
 from datetime import datetime
 
 import logging
@@ -24,17 +24,17 @@ def index(request):
         return render(request, 'index.html')
 
 
-def getHistoricalBinanceData(request):
+def get_historical_binance_data(request):
     client = BinanceClient()
     symbol = request.GET['symbol']
     interval = request.GET['interval']
     date_start = datetime.strptime(request.GET['date_start'], "%a, %d %b %Y %H:%M:%S %Z")
-    data = client.GetHistoricalDataWithInterval(symbol, interval, date_start)
+    data = client.get_historical_data_with_interval(symbol, interval, date_start)
     li = [{"open_time": record[0], "open": record[1]} for record in data]
     return JsonResponse(li, safe=False)
 
 
-def getHistoricalPoloniexData(request):
+def get_historical_poloniex_data(request):
     client = PoloniexClient()
     symbol = request.GET['symbol']
     interval = _map_intervals(request.GET['interval'])
